Rename (partially) previous Form
