+++ packages/sdk-components-react-remix/src/components.ts
@@ -1,4 +1,5 @@
 export { Body } from "./body";
 export { Link } from "./link";
 export { RichTextLink } from "./rich-text-link";
-export { Form } from "./form";
+// To not break compatibility with existing forms
+export { ServerForm as Form } from "./server-form";
