+++ packages/sdk-components-react-remix/src/metas.ts
@@ -1,4 +1,4 @@
 export { meta as Body } from "./body.ws";
 export { meta as Link } from "./link.ws";
 export { meta as RichTextLink } from "./rich-text-link.ws";
-export { meta as Form } from "./form.ws";
+export { meta as Form } from "./server-form.ws";
