+++ packages/sdk-components-react-remix/src/props.ts
@@ -1,4 +1,4 @@
 export { propsMeta as Body } from "./body.ws";
 export { propsMeta as Link } from "./link.ws";
 export { propsMeta as RichTextLink } from "./rich-text-link.ws";
-export { propsMeta as Form } from "./form.ws";
+export { propsMeta as Form } from "./server-form.ws";
